remove flags for now
We can always add the field back. I think it's best to remove them
for now cause we'll want to carefully examine how the field is handled
everywhere if it's re-introduced, instead of just leaving any 0 defaults
or copied values, since that might not be correct for a potential future
meaning for flags.